+++ SkyFloatingLabelTextFieldObjectiveCExample/Pods/SkyFloatingLabelTextField/Sources/SkyFloatingLabelTextField.swift
@@ -1,10 +1,13 @@
-//  Copyright 2016 Skyscanner Ltd
+//  Copyright 2016-2017 Skyscanner Ltd
 //
-//  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
+//  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance 
+//  with the License. You may obtain a copy of the License at
 //
 //  http://www.apache.org/licenses/LICENSE-2.0
 //
-//  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS /Users/gergelyorosz/Documents/Projects/Skyscanner/SkyFloatingLabelTextField/SkyFloatingLabelTextField/SkyFloatingLabelTextFieldTests/SkyFloatingLabelTextFieldTests.swiftOF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+//  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed 
+//  on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License 
+//  for the specific language governing permissions and limitations under the License.
 
 import UIKit
 
@@ -12,591 +15,594 @@ import UIKit
  A beautiful and flexible textfield implementation with support for title label, error message and placeholder.
  */
 @IBDesignable
-public class SkyFloatingLabelTextField: UITextField {
-    
+open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this type_body_length
+    /**
+     A Boolean value that determines if the language displayed is LTR. 
+     Default value set automatically from the application language settings.
+     */
+    open var isLTRLanguage = UIApplication.shared.userInterfaceLayoutDirection == .leftToRight {
+        didSet {
+           updateTextAligment()
+        }
+    }
+
+    fileprivate func updateTextAligment() {
+        if isLTRLanguage {
+            textAlignment = .left
+            titleLabel.textAlignment = .left
+        } else {
+            textAlignment = .right
+            titleLabel.textAlignment = .right
+        }
+    }
+
     // MARK: Animation timing
-    
+
     /// The value of the title appearing duration
-    public var titleFadeInDuration:NSTimeInterval = 0.2
+    dynamic open var titleFadeInDuration: TimeInterval = 0.2
     /// The value of the title disappearing duration
-    public var titleFadeOutDuration:NSTimeInterval = 0.3
-    
+    dynamic open var titleFadeOutDuration: TimeInterval = 0.3
+
     // MARK: Colors
-    
-    private var cachedTextColor:UIColor?
-    
+
+    fileprivate var cachedTextColor: UIColor?
+
     /// A UIColor value that determines the text color of the editable text
     @IBInspectable
-    override public var textColor:UIColor? {
+    override dynamic open var textColor: UIColor? {
         set {
-            self.cachedTextColor = newValue
-            self.updateControl(false)
+            cachedTextColor = newValue
+            updateControl(false)
         }
         get {
             return cachedTextColor
         }
     }
-    
+
     /// A UIColor value that determines text color of the placeholder label
-    @IBInspectable public var placeholderColor:UIColor = UIColor.lightGrayColor() {
+    @IBInspectable dynamic open var placeholderColor: UIColor = UIColor.lightGray {
         didSet {
-            self.updatePlaceholder()
+            updatePlaceholder()
         }
     }
 
     /// A UIColor value that determines text color of the placeholder label
-    @IBInspectable public var placeholderFont:UIFont? {
+    dynamic open var placeholderFont: UIFont? {
         didSet {
-            self.updatePlaceholder()
+            updatePlaceholder()
         }
     }
-    
-    private func updatePlaceholder() {
-        if let
-            placeholder = self.placeholder,
-            font = self.placeholderFont ?? self.font {
-                self.attributedPlaceholder = NSAttributedString(string: placeholder, attributes: [NSForegroundColorAttributeName:placeholderColor,
-                    NSFontAttributeName: font])
+
+    fileprivate func updatePlaceholder() {
+        if let placeholder = placeholder, let font = placeholderFont ?? font {
+                attributedPlaceholder = NSAttributedString(
+                    string: placeholder,
+                    attributes: [NSForegroundColorAttributeName: placeholderColor, NSFontAttributeName: font]
+                )
         }
     }
 
     /// A UIColor value that determines the text color of the title label when in the normal state
-    @IBInspectable public var titleColor:UIColor = UIColor.grayColor() {
+    @IBInspectable dynamic open var titleColor: UIColor = .gray {
         didSet {
-            self.updateTitleColor()
+            updateTitleColor()
         }
     }
-    
+
     /// A UIColor value that determines the color of the bottom line when in the normal state
-    @IBInspectable public var lineColor:UIColor = UIColor.lightGrayColor() {
+    @IBInspectable dynamic open var lineColor: UIColor = .lightGray {
         didSet {
-            self.updateLineView()
+            updateLineView()
         }
     }
-    
-    /// A UIColor value that determines the color used for the title label and the line when the error message is not `nil`
-    @IBInspectable public var errorColor:UIColor = UIColor.redColor() {
+
+    /// A UIColor value that determines the color used for the title label and line when the error message is not `nil`
+    @IBInspectable dynamic open var errorColor: UIColor = .red {
         didSet {
-            self.updateColors()
+            updateColors()
         }
     }
-    
+
     /// A UIColor value that determines the text color of the title label when editing
-    @IBInspectable public var selectedTitleColor:UIColor = UIColor.blueColor() {
+    @IBInspectable dynamic open var selectedTitleColor: UIColor = .blue {
         didSet {
-            self.updateTitleColor()
+            updateTitleColor()
         }
     }
-    
+
     /// A UIColor value that determines the color of the line in a selected state
-    @IBInspectable public var selectedLineColor:UIColor = UIColor.blackColor() {
+    @IBInspectable dynamic open var selectedLineColor: UIColor = .black {
         didSet {
-            self.updateLineView()
+            updateLineView()
         }
     }
-    
+
     // MARK: Line height
-    
+
     /// A CGFloat value that determines the height for the bottom line when the control is in the normal state
-    @IBInspectable public var lineHeight:CGFloat = 0.5 {
+    @IBInspectable dynamic open var lineHeight: CGFloat = 0.5 {
         didSet {
-            self.updateLineView()
-            self.setNeedsDisplay()
+            updateLineView()
+            setNeedsDisplay()
         }
     }
-    
+
     /// A CGFloat value that determines the height for the bottom line when the control is in a selected state
-    @IBInspectable public var selectedLineHeight:CGFloat = 1.0 {
+    @IBInspectable dynamic open var selectedLineHeight: CGFloat = 1.0 {
         didSet {
-            self.updateLineView()
-            self.setNeedsDisplay()
+            updateLineView()
+            setNeedsDisplay()
         }
     }
-    
+
     // MARK: View components
-    
+
     /// The internal `UIView` to display the line below the text input.
-    public var lineView:UIView!
-    
-    /// The internal `UILabel` that displays the selected, deselected title or the error message based on the current state.
-    public var titleLabel:UILabel!
-    
+    open var lineView: UIView!
+
+    /// The internal `UILabel` that displays the selected, deselected title or error message based on the current state.
+    open var titleLabel: UILabel!
+
     // MARK: Properties
-    
+
     /**
-    The formatter to use before displaying content in the title label. This can be the `title`, `selectedTitle` or the `errorMessage`.
+    The formatter used before displaying content in the title label. 
+    This can be the `title`, `selectedTitle` or the `errorMessage`.
     The default implementation converts the text to uppercase.
     */
-    public var titleFormatter:(String -> String) = { (text:String) -> String in
-        return text.uppercaseString
+    open var titleFormatter: ((String) -> String) = { (text: String) -> String in
+        return text.uppercased()
     }
-    
+
     /**
      Identifies whether the text object should hide the text being entered.
      */
-    override public var secureTextEntry:Bool {
+    override open var isSecureTextEntry: Bool {
         set {
-            super.secureTextEntry = newValue
-            self.fixCaretPosition()
+            super.isSecureTextEntry = newValue
+            fixCaretPosition()
         }
         get {
-            return super.secureTextEntry
+            return super.isSecureTextEntry
         }
     }
-    
+
     /// A String value for the error message to display.
-    public var errorMessage:String? {
+    open var errorMessage: String? {
         didSet {
-            self.updateControl(true)
+            updateControl(true)
         }
     }
-    
-    /// A Boolean value that determines whether the receiver discards `errorMessage` when the text input is changed.
-    public var discardsErrorMessageOnTextChange:Bool = true
-    
+
     /// The backing property for the highlighted property
-    private var _highlighted = false
-    
-    /// A Boolean value that determines whether the receiver is highlighted. When changing this value, highlighting will be done with animation
-    override public var highlighted:Bool {
+    fileprivate var _highlighted = false
+
+    /**
+     A Boolean value that determines whether the receiver is highlighted.
+     When changing this value, highlighting will be done with animation
+     */
+    override open var isHighlighted: Bool {
         get {
             return _highlighted
         }
         set {
-            self.setHighlighted(_highlighted, animated: true)
-        }
-    }
-    
-    /**
-     Sets the highlighted state with specifying whether this state change should be animated
-     
-     - parameter highlighted: Whether the field should be highlighted
-     - parameter animated: Whether the change in highlighting should be animated
-     */
-    public func setHighlighted(highlighted:Bool, animated:Bool) {
-        _highlighted = highlighted
-        if(self.highlighted) {
-            self.updateTitleColor()
-            _titleVisible = true
-            self.updateTitleVisibility(animated, animateFromCurrentState: true)
-            self.updateLineView()
-        } else {
-            if(animated) {
-                // Performing fading out after a short timeout to make sure the title previously faded in all the way
-                let time = dispatch_time(DISPATCH_TIME_NOW, Int64(self.titleFadeInDuration * Double(NSEC_PER_SEC)))
-                dispatch_after(time, dispatch_get_main_queue(), { () -> Void in
-                    self.fadeOutHighlightedWithAnimated(true)
-                })
-            } else {
-                self.fadeOutHighlightedWithAnimated(false)
-            }
+            _highlighted = newValue
+            updateTitleColor()
+            updateLineView()
         }
     }
 
     /// A Boolean value that determines whether the textfield is being edited or is selected.
-    public var editingOrSelected:Bool {
-        get {
-            return super.editing || self.selected;
-        }
+    open var editingOrSelected: Bool {
+        return super.isEditing || isSelected
     }
-    
+
     /// A Boolean value that determines whether the receiver has an error message.
-    public var hasErrorMessage:Bool {
-        get {
-            return self.errorMessage != nil
-        }
+    open var hasErrorMessage: Bool {
+        return errorMessage != nil && errorMessage != ""
     }
-    
-    private var _titleVisible:Bool = false
-    private var _renderingInInterfaceBuilder:Bool = false
-    
-    /// A Boolean value determining whether the title field is shown
-    public var titleVisible:Bool {
-        get {
-            return _titleVisible
-        }
-    }
-    private func setTitleVisibile(titleVisible:Bool, animated:Bool = false) {
-        if titleVisible != _titleVisible {
-            _titleVisible = titleVisible
-            self.updateTitleVisibility(animated)
-        }
-    }
-    
+
+    fileprivate var _renderingInInterfaceBuilder: Bool = false
+
     /// The text content of the textfield
     @IBInspectable
-    override public var text:String? {
+    override open var text: String? {
         didSet {
-            self.resetErrorMessageIfPresent()
-            self.updateControl(false)
+            updateControl(false)
         }
     }
-    
+
     /**
      The String to display when the input field is empty.
      The placeholder can also appear in the title label when both `title` `selectedTitle` and are `nil`.
      */
     @IBInspectable
-    override public var placeholder:String? {
+    override open var placeholder: String? {
         didSet {
-            self.setNeedsDisplay()
-            self.updateTitleLabel()
+            setNeedsDisplay()
+            updatePlaceholder()
+            updateTitleLabel()
         }
     }
-    
+
     /// The String to display when the textfield is editing and the input is not empty.
-    @IBInspectable public var selectedTitle:String? {
+    @IBInspectable open var selectedTitle: String? {
         didSet {
-            self.updateControl()
+            updateControl()
         }
     }
-    
+
     /// The String to display when the textfield is not editing and the input is not empty.
-    @IBInspectable public var title:String? {
+    @IBInspectable open var title: String? {
         didSet {
-            self.updateControl()
+            updateControl()
         }
     }
-    
+
     // Determines whether the field is selected. When selected, the title floats above the textbox.
-    public override var selected:Bool {
+    open override var isSelected: Bool {
         didSet {
-            self.updateControl(true)
+            updateControl(true)
         }
     }
-    
+
     // MARK: - Initializers
-    
+
     /**
     Initializes the control
     - parameter frame the frame of the control
     */
     override public init(frame: CGRect) {
         super.init(frame: frame)
-        self.init_SkyFloatingLabelTextField()
+        init_SkyFloatingLabelTextField()
     }
-    
+
     /**
      Intialzies the control by deserializing it
      - parameter coder the object to deserialize the control from
      */
     required public init?(coder aDecoder: NSCoder) {
         super.init(coder: aDecoder)
-        self.init_SkyFloatingLabelTextField()
+        init_SkyFloatingLabelTextField()
     }
-    
-    private final func init_SkyFloatingLabelTextField() {
-        self.borderStyle = .None
-        self.createTitleLabel()
-        self.createLineView()
-        self.updateColors()
-        self.addEditingChangedObserver()
+
+    fileprivate final func init_SkyFloatingLabelTextField() {
+        borderStyle = .none
+        createTitleLabel()
+        createLineView()
+        updateColors()
+        addEditingChangedObserver()
+        updateTextAligment()
     }
-    
-    private func addEditingChangedObserver() {
-        self.addTarget(self, action: Selector("editingChanged"), forControlEvents: .EditingChanged)
+
+    fileprivate func addEditingChangedObserver() {
+        self.addTarget(self, action: #selector(SkyFloatingLabelTextField.editingChanged), for: .editingChanged)
     }
-    
+
     /**
      Invoked when the editing state of the textfield changes. Override to respond to this change.
      */
-    public func editingChanged() {
-        resetErrorMessageIfPresent()
+    open func editingChanged() {
         updateControl(true)
         updateTitleLabel(true)
     }
-    
+
     // MARK: create components
-    
-    private func createTitleLabel() {
+
+    fileprivate func createTitleLabel() {
         let titleLabel = UILabel()
-        titleLabel.autoresizingMask = [.FlexibleWidth, .FlexibleHeight]
-        titleLabel.font = UIFont.systemFontOfSize(13)
+        titleLabel.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        titleLabel.font = .systemFont(ofSize: 13)
         titleLabel.alpha = 0.0
-        titleLabel.textColor = self.titleColor
-        self.addSubview(titleLabel)
+        titleLabel.textColor = titleColor
+
+        addSubview(titleLabel)
         self.titleLabel = titleLabel
     }
-    
-    private func createLineView() {
-        
-        if self.lineView == nil {
+
+    fileprivate func createLineView() {
+
+        if lineView == nil {
             let lineView = UIView()
-            lineView.userInteractionEnabled = false
+            lineView.isUserInteractionEnabled = false
             self.lineView = lineView
-            self.configureDefaultLineHeight()
+            configureDefaultLineHeight()
         }
-        lineView.autoresizingMask = [.FlexibleWidth, .FlexibleTopMargin]
-        self.addSubview(lineView)
+
+        lineView.autoresizingMask = [.flexibleWidth, .flexibleTopMargin]
+        addSubview(lineView)
     }
-    
-    private func configureDefaultLineHeight() {
-        let onePixel:CGFloat = 1.0 / UIScreen.mainScreen().scale
-        self.lineHeight = 2.0 * onePixel
-        self.selectedLineHeight = 2.0 * self.lineHeight
+
+    fileprivate func configureDefaultLineHeight() {
+        let onePixel: CGFloat = 1.0 / UIScreen.main.scale
+        lineHeight = 2.0 * onePixel
+        selectedLineHeight = 2.0 * self.lineHeight
     }
-    
+
     // MARK: Responder handling
-    
+
     /**
      Attempt the control to become the first responder
      - returns: True when successfull becoming the first responder
     */
-    override public func becomeFirstResponder() -> Bool {
+    @discardableResult
+    override open func becomeFirstResponder() -> Bool {
         let result = super.becomeFirstResponder()
-        self.updateControl(true)
+        updateControl(true)
         return result
     }
-    
+
     /**
      Attempt the control to resign being the first responder
      - returns: True when successfull resigning being the first responder
      */
-    override public func resignFirstResponder() -> Bool {
+    @discardableResult
+    override open func resignFirstResponder() -> Bool {
         let result =  super.resignFirstResponder()
-        self.updateControl(true)
+        updateControl(true)
         return result
     }
-    
+
     // MARK: - View updates
-    
-    private func updateControl(animated:Bool = false) {
-        self.updateColors()
-        self.updateLineView()
-        self.updateTitleLabel(animated)
+
+    fileprivate func updateControl(_ animated: Bool = false) {
+        updateColors()
+        updateLineView()
+        updateTitleLabel(animated)
     }
-    
-    private func updateLineView() {
-        if let lineView = self.lineView {
-            lineView.frame = self.lineViewRectForBounds(self.bounds, editing: self.editingOrSelected)
+
+    fileprivate func updateLineView() {
+        if let lineView = lineView {
+            lineView.frame = lineViewRectForBounds(bounds, editing: editingOrSelected)
         }
-        self.updateLineColor()
+        updateLineColor()
     }
-    
+
     // MARK: - Color updates
-    
+
     /// Update the colors for the control. Override to customize colors.
-    public func updateColors() {
-        self.updateLineColor()
-        self.updateTitleColor()
-        self.updateTextColor()
+    open func updateColors() {
+        updateLineColor()
+        updateTitleColor()
+        updateTextColor()
     }
-    
-    private func updateLineColor() {
-        if self.hasErrorMessage {
-            self.lineView.backgroundColor = self.errorColor
+
+    fileprivate func updateLineColor() {
+        if hasErrorMessage {
+            lineView.backgroundColor = errorColor
         } else {
-            self.lineView.backgroundColor = self.editingOrSelected ? self.selectedLineColor : self.lineColor
+            lineView.backgroundColor = editingOrSelected ? selectedLineColor : lineColor
         }
     }
-    
-    private func updateTitleColor() {
-        if self.hasErrorMessage {
-            self.titleLabel.textColor = self.errorColor
+
+    fileprivate func updateTitleColor() {
+        if hasErrorMessage {
+            titleLabel.textColor = errorColor
         } else {
-            if self.editingOrSelected || self.highlighted {
-                self.titleLabel.textColor = self.selectedTitleColor
+            if editingOrSelected || isHighlighted {
+                titleLabel.textColor = selectedTitleColor
             } else {
-                self.titleLabel.textColor = self.titleColor
+                titleLabel.textColor = titleColor
             }
         }
     }
-    
-    private func updateTextColor() {
-        if self.hasErrorMessage {
-            super.textColor = self.errorColor
+
+    fileprivate func updateTextColor() {
+        if hasErrorMessage {
+            super.textColor = errorColor
         } else {
-            super.textColor = self.cachedTextColor
+            super.textColor = cachedTextColor
         }
     }
-    
+
     // MARK: - Title handling
-    
-    private func updateTitleLabel(animated:Bool = false) {
-        
-        var titleText:String? = nil
-        if self.hasErrorMessage {
-            titleText = self.titleFormatter(errorMessage!)
+
+    fileprivate func updateTitleLabel(_ animated: Bool = false) {
+
+        var titleText: String? = nil
+        if hasErrorMessage {
+            titleText = titleFormatter(errorMessage!)
         } else {
-            if self.editingOrSelected {
-                titleText = self.selectedTitleOrTitlePlaceholder()
+            if editingOrSelected {
+                titleText = selectedTitleOrTitlePlaceholder()
                 if titleText == nil {
-                    titleText = self.titleOrPlaceholder()
+                    titleText = titleOrPlaceholder()
                 }
             } else {
-                titleText = self.titleOrPlaceholder()
+                titleText = titleOrPlaceholder()
             }
         }
-        self.titleLabel.text = titleText
-        
-        let titleVisible = (titleText != nil) && self.hasText()
-        self.setTitleVisibile(titleVisible, animated: animated)
+        titleLabel.text = titleText
+
+        updateTitleVisibility(animated)
     }
-    
-    private func updateTitleVisibility(animated:Bool = false, animateFromCurrentState:Bool = false) {
-        let alpha:CGFloat = _titleVisible ? 1.0 : 0.0
-        let frame:CGRect = self.titleLabelRectForBounds(self.bounds, editing: _titleVisible)
+
+    fileprivate var _titleVisible = false
+
+    /*
+    *   Set this value to make the title visible
+    */
+    open func setTitleVisible(
+        _ titleVisible: Bool,
+        animated: Bool = false,
+        animationCompletion: ((_ completed: Bool) -> Void)? = nil
+    ) {
+        if _titleVisible == titleVisible {
+            return
+        }
+        _titleVisible = titleVisible
+        updateTitleColor()
+        updateTitleVisibility(animated, completion: animationCompletion)
+    }
+
+    /**
+     Returns whether the title is being displayed on the control.
+     - returns: True if the title is displayed on the control, false otherwise.
+     */
+    open func isTitleVisible() -> Bool {
+        return hasText || hasErrorMessage || _titleVisible
+    }
+
+    fileprivate func updateTitleVisibility(_ animated: Bool = false, completion: ((_ completed: Bool) -> Void)? = nil) {
+        let alpha: CGFloat = isTitleVisible() ? 1.0 : 0.0
+        let frame: CGRect = titleLabelRectForBounds(bounds, editing: isTitleVisible())
         let updateBlock = { () -> Void in
             self.titleLabel.alpha = alpha
             self.titleLabel.frame = frame
         }
         if animated {
-            var animationOptions:UIViewAnimationOptions = .CurveEaseOut;
-            if(animateFromCurrentState) {
-                animationOptions = [.BeginFromCurrentState, .CurveEaseOut]
-            }
-            let duration = _titleVisible ? titleFadeInDuration : titleFadeOutDuration
-            
-            UIView.animateWithDuration(duration, delay: 0, options: animationOptions, animations: { () -> Void in
+            let animationOptions: UIViewAnimationOptions = .curveEaseOut
+            let duration = isTitleVisible() ? titleFadeInDuration : titleFadeOutDuration
+            UIView.animate(withDuration: duration, delay: 0, options: animationOptions, animations: { () -> Void in
                 updateBlock()
-                }, completion: nil)
+                }, completion: completion)
         } else {
             updateBlock()
+            completion?(true)
         }
     }
-    
+
     // MARK: - UITextField text/placeholder positioning overrides
-    
-    /** 
+
+    /**
     Calculate the rectangle for the textfield when it is not being edited
     - parameter bounds: The current bounds of the field
     - returns: The rectangle that the textfield should render in
     */
-    override public func textRectForBounds(bounds: CGRect) -> CGRect {
-        super.textRectForBounds(bounds)
-        let titleHeight = self.titleHeight()
-        let lineHeight = self.selectedLineHeight
-        let rect = CGRectMake(0, titleHeight, bounds.size.width, bounds.size.height - titleHeight - lineHeight)
+    override open func textRect(forBounds bounds: CGRect) -> CGRect {
+        super.textRect(forBounds: bounds)
+        let rect = CGRect(
+            x: 0,
+            y: titleHeight(),
+            width: bounds.size.width,
+            height: bounds.size.height - titleHeight() - selectedLineHeight
+        )
         return rect
     }
-    
+
     /**
      Calculate the rectangle for the textfield when it is being edited
      - parameter bounds: The current bounds of the field
      - returns: The rectangle that the textfield should render in
      */
-    override public func editingRectForBounds(bounds: CGRect) -> CGRect {
-        let titleHeight = self.titleHeight()
-        let lineHeight = self.selectedLineHeight
-        let rect = CGRectMake(0, titleHeight, bounds.size.width, bounds.size.height - titleHeight - lineHeight)
+    override open func editingRect(forBounds bounds: CGRect) -> CGRect {
+        let rect = CGRect(
+            x: 0,
+            y: titleHeight(),
+            width: bounds.size.width,
+            height: bounds.size.height - titleHeight() - selectedLineHeight
+        )
         return rect
     }
-    
+
     /**
      Calculate the rectangle for the placeholder
      - parameter bounds: The current bounds of the placeholder
      - returns: The rectangle that the placeholder should render in
      */
-    override public func placeholderRectForBounds(bounds: CGRect) -> CGRect {
-        let titleHeight = self.titleHeight()
-        let lineHeight = self.selectedLineHeight
-        let rect = CGRectMake(0, titleHeight, bounds.size.width, bounds.size.height - titleHeight - lineHeight)
+    override open func placeholderRect(forBounds bounds: CGRect) -> CGRect {
+        let rect = CGRect(
+            x: 0,
+            y: titleHeight(),
+            width: bounds.size.width,
+            height: bounds.size.height - titleHeight() - selectedLineHeight
+        )
         return rect
     }
-    
+
     // MARK: - Positioning Overrides
-    
+
     /**
     Calculate the bounds for the title label. Override to create a custom size title field.
     - parameter bounds: The current bounds of the title
     - parameter editing: True if the control is selected or highlighted
     - returns: The rectangle that the title label should render in
     */
-    public func titleLabelRectForBounds(bounds:CGRect, editing:Bool) -> CGRect {
-        let titleHeight = self.titleHeight()
+    open func titleLabelRectForBounds(_ bounds: CGRect, editing: Bool) -> CGRect {
         if editing {
-            return CGRectMake(0, 0, bounds.size.width, titleHeight)
+            return CGRect(x: 0, y: 0, width: bounds.size.width, height: titleHeight())
         }
-        return CGRectMake(0, titleHeight, bounds.size.width, titleHeight)
+        return CGRect(x: 0, y: titleHeight(), width: bounds.size.width, height: titleHeight())
     }
 
     /**
-     Calculate the bounds for the bottom line of the control. Override to create a custom size bottom line in the textbox.
+     Calculate the bounds for the bottom line of the control. 
+     Override to create a custom size bottom line in the textbox.
      - parameter bounds: The current bounds of the line
      - parameter editing: True if the control is selected or highlighted
      - returns: The rectangle that the line bar should render in
      */
-    public func lineViewRectForBounds(bounds:CGRect, editing:Bool) -> CGRect {
-        let lineHeight:CGFloat = editing ? CGFloat(self.selectedLineHeight) : CGFloat(self.lineHeight)
-        return CGRectMake(0, bounds.size.height - lineHeight, bounds.size.width, lineHeight);
+    open func lineViewRectForBounds(_ bounds: CGRect, editing: Bool) -> CGRect {
+        let height = editing ? selectedLineHeight : lineHeight
+        return CGRect(x: 0, y: bounds.size.height - height, width: bounds.size.width, height: height)
     }
-    
+
     /**
      Calculate the height of the title label.
      -returns: the calculated height of the title label. Override to size the title with a different height
      */
-    public func titleHeight() -> CGFloat {
-        if let titleLabel = self.titleLabel,
-            font = titleLabel.font {
-                return font.lineHeight
+    open func titleHeight() -> CGFloat {
+        if let titleLabel = titleLabel,
+            let font = titleLabel.font {
+            return font.lineHeight
         }
         return 15.0
     }
-    
+
     /**
      Calcualte the height of the textfield.
      -returns: the calculated height of the textfield. Override to size the textfield with a different height
      */
-    public func textHeight() -> CGFloat {
+    open func textHeight() -> CGFloat {
         return self.font!.lineHeight + 7.0
     }
-    
+
     // MARK: - Layout
-    
+
     /// Invoked when the interface builder renders the control
-    override public func prepareForInterfaceBuilder() {
-        super.prepareForInterfaceBuilder()
-        self.selected = true
+    override open func prepareForInterfaceBuilder() {
+        if #available(iOS 8.0, *) {
+            super.prepareForInterfaceBuilder()
+        }
+
+        borderStyle = .none
+
+        isSelected = true
         _renderingInInterfaceBuilder = true
-        self.updateControl(false)
-        self.invalidateIntrinsicContentSize()
+        updateControl(false)
+        invalidateIntrinsicContentSize()
     }
-    
+
     /// Invoked by layoutIfNeeded automatically
-    override public func layoutSubviews() {
+    override open func layoutSubviews() {
         super.layoutSubviews()
-        
-        self.titleLabel.frame = self.titleLabelRectForBounds(self.bounds, editing: self.hasText() || _renderingInInterfaceBuilder)
-        self.lineView.frame = self.lineViewRectForBounds(self.bounds, editing: self.editingOrSelected || _renderingInInterfaceBuilder)
+
+        titleLabel.frame = titleLabelRectForBounds(bounds, editing: isTitleVisible() || _renderingInInterfaceBuilder)
+        lineView.frame = lineViewRectForBounds(bounds, editing: editingOrSelected || _renderingInInterfaceBuilder)
     }
-    
+
     /**
      Calculate the content size for auto layout
-     
+
      - returns: the content size to be used for auto layout
      */
-    override public func intrinsicContentSize() -> CGSize {
-        return CGSizeMake(self.bounds.size.width, self.titleHeight() + self.textHeight())
+    override open var intrinsicContentSize: CGSize {
+        return CGSize(width: bounds.size.width, height: titleHeight() + textHeight())
     }
-    
+
     // MARK: - Helpers
-    
-    private func fadeOutHighlightedWithAnimated(animated: Bool) {
-        if(!self.hasText()) {
-            self.updateTitleColor()
-            _titleVisible = false
-            self.updateTitleVisibility(animated, animateFromCurrentState: true)
-            self.updateLineView()
+
+    fileprivate func titleOrPlaceholder() -> String? {
+        guard let title = title ?? placeholder else {
+            return nil
         }
+        return titleFormatter(title)
     }
-    
-    private func resetErrorMessageIfPresent() {
-        if self.hasErrorMessage && discardsErrorMessageOnTextChange {
-            self.errorMessage = nil
+
+    fileprivate func selectedTitleOrTitlePlaceholder() -> String? {
+        guard let title = selectedTitle ?? title ?? placeholder else {
+            return nil
         }
+        return titleFormatter(title)
     }
-    
-    private func titleOrPlaceholder() -> String? {
-        if let title = self.title ?? self.placeholder {
-            return self.titleFormatter(title)
-        }
-        return nil
-    }
-    
-    private func selectedTitleOrTitlePlaceholder() -> String? {
-        if let title = self.selectedTitle ?? self.title ?? self.placeholder {
-            return self.titleFormatter(title)
-        }
-        return nil
-    }
-}
+} // swiftlint:disable:this file_length