+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Master
 ------
 
+v3.2.0
+------
+* Fixed editing rect size when clear button is present to no overlap [#148](https://github.com/Skyscanner/SkyFloatingLabelTextField/pull/148)
+* Added `titleFont` as open var to set another font in `titleLabel` [#141](https://github.com/Skyscanner/SkyFloatingLabelTextField/pull/141). Thanks to [@rabenson](https://github.com/rabenson) and [@rubenadlm](https://github.com/rubenadlm).
+
 v3.1.0
 ------